gt 변수를 long 타입으로 변환하는 변경사항을 추가했습니다.

+++ models/dpt.py
@@ -991,7 +991,7 @@ class BaseModel(L.LightningModule):
         #carbon = carbon.to(dtype=torch.long)
         #gt = gt.to(dtype=torch.long)
         gt_pred, carbon_pred = self(x)
-        
+        gt = gt.long()
         gt_loss = F.cross_entropy(gt_pred, gt)
         carbon_pred = carbon_pred.squeeze(1)
         carbon_loss = F.mse_loss(carbon_pred, carbon)
